requirements.txt
backports.functools-lru-cache==1.6.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
cssselect==1.1.0
DateTime==4.3
dnspython==1.16.0
elasticsearch==7.5.1
enum34==1.1.9
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
geographiclib==1.50
geopy==1.21.0
http-ece==1.1.0
httpagentparser==1.9.0
idna==2.9
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
markdown2==2.3.8
MarkupSafe==1.1.1
Pillow==6.2.2
psycopg2-binary==2.8.4
py-vapid==1.7.0
pycparser==2.19
PyJWT==1.7.1
pyquery==1.4.1
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
pywebpush==1.10.1
readtime==1.1.1
redis==3.4.1
regex==2020.2.20
requests==2.23.0
retinasdk==1.0.0
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
urllib3==1.25.8
webptools==0.0.2
Werkzeug==1.0.0
zope.interface==4.7.1
